requirements.txt
langchain
ipykernel
python-dotenv
langchain_openai
langchain-huggingface
chromadb
beautifulsoup4
langchain-community
langchain_classic
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.41.2
numpy==1.26.4
langchain-chroma
langchain-text-splitters
pypdf
numexpr
wikipedia